some changes

+++ src/main/java/com/example/asthmatracker/models/Spirometry.java
@@ -11,6 +11,6 @@ import java.time.LocalDateTime;
 public class Spirometry {
     private Integer id;
     private Integer patient_id;
-    private BigDecimal result;
+    private Integer result;
     private LocalDateTime date_time;
 }
